Change: Restruct package by moving components into 'components' sub-package for more clear structure.
