+++ llyra/components/logs/utils/funcs.py
